+++ src/lib/util.h
@@ -23,14 +23,15 @@ extern void format(char *buf, char *fmt, ...);
 #define SMAX 1000
 #define TMAX 1000000
 
-#define member(c, set) strchr(set, c)
-#define error(...) { printf(__VA_ARGS__); assert(0); }
+#define BIT(var, pos) ((var) & (1<<(pos)))
 #define eq(x,y) (strcmp((x),(y))==0)
 #define debug(...) { if (isDebug) printf(__VA_ARGS__); }
 #define min(x,y) (((x)<(y)) ? (x) : (y))
 #define max(x,y) (((x)>(y)) ? (x) : (y))
+#define member(c, set) strchr(set, c)
+#define error(...) { printf(__VA_ARGS__); assert(0); }
 
-extern char isDebug;
+extern char isDebug, *argOut;
 extern char isFlag[];
 
 #endif