+++ src/ma/0.5/macro.c
@@ -8,7 +8,6 @@ Pair macroList[] = {
   {"puti", "@${1}\nD=M\n@0\nswi"},
   {"putc", "@${1}\nD=M\n@1\nswi"},
   {"puts", "@${1}\nD=A\n@3\nswi"},
-  {"puttime", "@15\nswi"},
   // float
   {"putf", "@18\nswi"},
   {"setf", "@${1}\nD=A\n@17\nswi"},
@@ -18,15 +17,11 @@ Pair macroList[] = {
   {"divf", "@${1}\nD=A\n@22\nswi"},
   // control
   {"set",  "@${3}\nD=M\n@${1}\nM=D"},   // .set  i = t1
-  {"setc", "@${3}\nD=A\n@${1}\nM=D"},   // .setc i = 5
+  {"seti", "@${3}\nD=A\n@${1}\nM=D"},   // .seti i = 5
   {"goto", "@${1}\n0;JMP"},             // goto L
   {"if",   "@${1}\nD=M\n@${3}\nD;JNE"}, // if e goto L
   {"ifnot","@${1}\nD=M\n@${3}\nD;JEQ"}, // ifnot e goto L
   {"op",   "@${3}\nD=M\n@${5}\nD=D${4}M\n@${1}\nM=D"}, // .op z = x + y
-  {"opc",  "@${3}\nD=M\n@${5}\nD=D${4}A\n@${1}\nM=D"}, // .op z = x + 1
-  // function
-  {"call",  "@${1}\ncall"},   // .call f
-  {"ret",   "ret"},         // .ret
 };
 
 Map macroMap;
@@ -44,34 +39,33 @@ int expand(char *line, char *code) {
   return 1;
 }
 
-void macroExpand(char *iFile, FILE *oF) {
+void macroExpand(char *iFile, char *oFile) {
   char line[SMAX];
   debug("====== macroExpand ============\n");
   FILE *iF = fopen(iFile, "r");
   if (iF == NULL) error("macroExpand: file %s not found!\n", iFile);
-  char code[TMAX];
-  sprintf(code, "// =========== iFile: %s ==============\n", iFile);
-  fwrite(code, strlen(code), 1, oF);
+  FILE *oF = fopen(oFile, "w");
   while (fgets(line, sizeof(line), iF)) {
+    char code[TMAX];
     int isExpand = expand(line, code);
     if (isExpand) { debug("%s", code); } else debug("%s", line);
     fwrite(code, strlen(code), 1, oF);
   }
   fclose(iF);
+  fclose(oF);
 }
 
-// run: ./ma iFile1, 2, .... -o oFile -d
+// run: ./ma <file> 
+// notice : <file> with no extension.
 int main(int argc, char *argv[]) {
   argHandle(argc, argv, 2, "./ma <file>\n");
 
-  mapNew(&macroMap, 59); mapAddAll(&macroMap, macroList, ARRAY_SIZE(macroList));
+  char *file = argv[1];
+  char iFile[SMAX], oFile[SMAX];
+  sprintf(iFile, "%s.mx", file);
+  sprintf(oFile, "%s.sx", file);
 
-  FILE *oF = fopen(argOut, "w");
-  for (int i=1; i<argc; i++) {
-    char *iFile = argv[i];
-    if (strcmp(argv[i], "-o") == 0) break;
-    macroExpand(iFile, oF);
-  }
-
+  mapNew(&macroMap, 37); mapAddAll(&macroMap, macroList, ARRAY_SIZE(macroList));
+  macroExpand(iFile, oFile);
   mapFree(&macroMap);
 }